requirements.txt
django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.3.1
Pillow
Faker
pyotp
openpyxl==3.1.2  # Для экспорта в Excel
faker==20.1.0